feat: shitload of changes

+++ model/data_models/capacity.py
@@ -3,6 +3,6 @@ from dataclasses import dataclass
 @dataclass
 class FrameCapacity:
     name: str
-    mass: int
+    mass: float
     disp: list[float]
     base_shear: list[float]
